requirements.txt
streamlit
langchain-core
langchain-coummunity
langchain-huggingface
langchain-text-splitters
langchain-chroma
python-dotenv
mysql-connector-python
streamlit-cookies-manager
snowflake-id
pytesseract
chromadb
pyMuPDF
Pillow
